update content urls
